requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
cryptography>=42.0
pycryptodome>=3.20
streamlit>=1.20.0
